Codeforces 1183 - C
downcasting error...

+++ codeforces/1183/C.cc
@@ -3,7 +3,7 @@ using namespace std;
 int query() {
   int k,n,a,b;
   scanf("%d%d%d%d",&k,&n,&a,&b);
-  int ret = ((k - 1LL*b*n) + (a-b-1)) / (a-b) - 1;
+  long long ret = (k-1LL*b*n + (a-b-1)) / (a-b) - 1;
   if(ret > n) return n;
   if(ret < 0) return -1;
   return ret;
